204: Move test from issue 198 to PersistentOrderedMapTest
