+++ core/commonTest/src/contract/map/PersistentHashMapTest.kt
@@ -6,9 +6,7 @@
 package tests.contract.map
 
 import kotlinx.collections.immutable.implementations.immutableMap.PersistentHashMap
-import kotlinx.collections.immutable.minus
 import kotlinx.collections.immutable.persistentHashMapOf
-import kotlinx.collections.immutable.persistentMapOf
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertTrue
@@ -33,44 +31,5 @@ class PersistentHashMapTest {
         assertTrue(map3.equals(builder))
         assertEquals(map3, map4.toMap())
         assertEquals(map3, map4)
-    }
-
-    /**
-     * Test from issue: https://github.com/Kotlin/kotlinx.collections.immutable/issues/198
-     */
-    @Test
-    fun `if the full collision is of size 3 and 2 of the keys is removed the remaining key must be promoted`() {
-        class ChosenHashCode(
-            private val hashCode: Int,
-            private val name: String,
-        ) {
-            override fun equals(other: Any?): Boolean {
-                return other is ChosenHashCode && (other.name == name)
-            }
-
-            override fun hashCode(): Int {
-                return hashCode
-            }
-
-            override fun toString(): String {
-                return name
-            }
-        }
-
-        val a = ChosenHashCode(123, "A")
-        val b = ChosenHashCode(123, "B")
-        val c = ChosenHashCode(123, "C")
-
-        val abc = persistentMapOf(
-            a to "x",
-            b to "y",
-            c to "z",
-        )
-
-        val minusAb = abc.minus(arrayOf(a, b))
-        val cOnly = persistentMapOf(c to "z")
-
-        assertEquals(minusAb.entries, cOnly.entries)
-        assertEquals(minusAb, cOnly)
     }
 }
